refactor: dedicated vitest project for compliance suite, drop env-var gating
Compliance registration moves out of fork_choice.test.ts into its own
test/spec/comptest/ file backed by a dedicated `comptest` vitest
project; the shared runner is extracted to utils/forkChoiceTestRunner.ts.

The regular spec projects exclude test/spec/comptest/** so
test:spec/test:spec:minimal can never pick the suite up, which removes
the need for RUN_FORK_CHOICE_COMPLIANCE: `pnpm test:comptest` simply
runs the comptest project. Zero-test guard (workspace config has
passWithNoTests: true) moves into the comptest entry file, now
unconditional.
