+++ packages/beacon-node/package.json
@@ -101,7 +101,7 @@
     "test:sim:blobs": "vitest run test/sim/4844-interop.test.ts",
     "download-spec-tests": "node --loader=ts-node/esm test/spec/downloadTests.ts",
     "download-comptests": "node --loader=ts-node/esm test/spec/downloadComptests.ts",
-    "test:comptest": "RUN_FORK_CHOICE_COMPLIANCE=1 vitest run --project spec-minimal test/spec/presets/fork_choice.test.ts -t fork_choice_compliance",
+    "test:comptest": "vitest run --project comptest",
     "test:spec:bls": "vitest run --project spec-minimal test/spec/bls/",
     "test:spec:general": "vitest run --project spec-minimal test/spec/general/",
     "test:spec:minimal": "vitest run --project spec-minimal test/spec/presets/",
